1. rename the file
