+++ src/main/java/interface_adaptors/servletDeleteResponse.java
@@ -1,4 +1,4 @@
-package interfaceAdaptors;
+package interface_adaptors;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -7,11 +7,12 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-// The servletLoginResponseUserNotExist class works as a presenter and pass user not exist message back to front-end.
-@WebServlet("/loginResponseUserNotExist")
-public class servletLoginResponseUserNotExist extends HttpServlet {
+// The servletDeleteResponse class works as a presenter and pass message back to front-end.
+@WebServlet("/deleteResponse")
+public class servletDeleteResponse extends HttpServlet {
     @Override
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        resp.getWriter().append("not_exist");
+        String status = (String) req.getAttribute("status");
+        resp.getWriter().append(status);
     }
 }
