+++ src/main/java/interface_adaptors/servletGetHistory.java
@@ -1,8 +1,8 @@
-package interfaceAdaptors;
+package interface_adaptors;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import database_connection.DatabaseRead;
-import post_reply_user.Post;
+import database_connection.*;
+import post_reply_user.*;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -12,10 +12,11 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.List;
 
-@WebServlet("/sendPostPresenter")
-public class servletSendPostPresenter extends HttpServlet {
+@WebServlet("/servletGetHistory")
+public class servletGetHistory extends HttpServlet{
     @Override
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        req.setCharacterEncoding("UTF-8");
         resp.setCharacterEncoding("UTF-8");
         resp.setContentType("application/json");
