+++ src/main/java/interface_adaptors/servletLoginResponseSuccess.java
@@ -1,4 +1,4 @@
-package interfaceAdaptors;
+package interface_adaptors;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -7,11 +7,11 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-// The servletLogin class works as a controller and pass variables to gateway.
-@WebServlet("/login")
-public class servletLogin extends HttpServlet {
+// The servletLoginResponseSuccess class works as a presenter and pass success message back to front-end.
+@WebServlet("/loginResponseSuccess")
+public class servletLoginResponseSuccess extends HttpServlet {
     @Override
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        req.getRequestDispatcher("/loginGateway").forward(req,resp);
+        resp.getWriter().append("success");
     }
 }
