+++ src/main/java/interface_adaptors/servletPostPostResponse.java
@@ -1,6 +1,4 @@
-package interfaceAdaptors;
-
-import database_connection.Database;
+package interface_adaptors;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -9,11 +7,10 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-// The servletSendReply class works as a controller and pass variables to gateway.
-@WebServlet("/listenSendReply")
-public class servletSendReply extends HttpServlet {
+// The servletPostPostResponse class works as a presenter and pass message back to front-end.
+@WebServlet("/postPostResponse")
+public class servletPostPostResponse extends HttpServlet {
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        req.setCharacterEncoding("UTF-8");
-        req.getRequestDispatcher("/sendReplyGateway").forward(req,resp);
+        resp.getWriter().append("success");
     }
 }
