+++ src/main/java/interface_adaptors/servletRegisterResponsePasswordInvalid.java
@@ -1,4 +1,4 @@
-package interfaceAdaptors;
+package interface_adaptors;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -6,11 +6,11 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-// The servletLikePostResponseLike class works as a presenter and pass like message back to front-end.
-@WebServlet("/likePostResponseLike")
-public class servletLikePostResponseLike extends HttpServlet {
-    @Override
+
+// The servletRegisterResponsePasswordInvalid class works as a presenter and pass incorrect message back to front-end.
+@WebServlet("/registerResponsePasswordInvalid")
+public class servletRegisterResponsePasswordInvalid extends HttpServlet {
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        resp.getWriter().append("succeed to like");
+        resp.getWriter().append("password not match");
     }
 }
