+++ src/main/java/interface_adaptors/servletSendPostPresenter.java
@@ -1,8 +1,8 @@
-package interfaceAdaptors;
+package interface_adaptors;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import database_connection.*;
-import post_reply_user.*;
+import database_connection.DatabaseRead;
+import post_reply_user.Post;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -12,11 +12,10 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.List;
 
-@WebServlet("/servletGetHistory")
-public class servletGetHistory extends HttpServlet{
+@WebServlet("/sendPostPresenter")
+public class servletSendPostPresenter extends HttpServlet {
     @Override
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        req.setCharacterEncoding("UTF-8");
         resp.setCharacterEncoding("UTF-8");
         resp.setContentType("application/json");
