+++ src/main/java/interface_adaptors/servletSendReplyResponse.java
@@ -1,4 +1,4 @@
-package interfaceAdaptors;
+package interface_adaptors;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -7,10 +7,12 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-// The servletRegisterResponsePasswordInvalid class works as a presenter and pass incorrect message back to front-end.
-@WebServlet("/registerResponsePasswordInvalid")
-public class servletRegisterResponsePasswordInvalid extends HttpServlet {
+// The servletSendReplyResponse class works as a presenter and pass message back to front-end.
+@WebServlet("/sendReplyResponse")
+public class servletSendReplyResponse extends HttpServlet {
+    @Override
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        resp.getWriter().append("password not match");
+        resp.setCharacterEncoding("UTF-8");
+        resp.getWriter().append("success");
     }
 }
